move skywalking_to_traces into pkg/translator
