+++ pkg/translator/skywalking/metadata.yaml
@@ -1,0 +1,3 @@
+status:
+  codeowners:
+    active: [JaredTan95]
