+++ pkg/translator/skywalking/skywalkingproto_to_traces_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package trace
+package skywalking
 
 import (
 	"strconv"
